fix: typescript cyclic import fixed
